WIP
An attempt to run `gdn server` directly as maximus via BPM, with all
setup occurring in a bpm pre_start script.

Got to the point where cgroups are an issue - can't mount cgroups in
pre_start as they won't be mounted in the bpm container's mount NS. Also
can't mount them in the bpm container as we are maximus and don't have
caps. wut do?

Signed-off-by: Claudia Beresford <cberesford@pivotal.io>